+++ src/components/chart/chart-functions.ts
@@ -2,7 +2,7 @@ import { addDays, eachDayOfInterval } from "date-fns";
 import { formatDate, parseDateString, maxDateStr } from "src/lib/date.ts";
 
 import type { Tally } from 'src/lib/api/tally.ts';
-import { TALLY_MEASURE_INFO } from "src/lib/tally.ts";
+import { TALLY_MEASURE_INFO, cmpTallies } from "src/lib/tally.ts";
 
 export interface TallyPoint {
   date: string;
@@ -25,56 +25,6 @@ export interface CompiledTallyPoint {
   total: {
     [measure in keyof typeof TALLY_MEASURE_INFO]: number;
   };
-}
-
-export function cmpTallies(a: TallyPoint | Tally, b: TallyPoint | Tally) {
-  return a.date < b.date ? -1 : a.date > b.date ? 1 : 0;
-}
-
-export function compileTallies(tallies: Tally[], overrideStartDate: string = null, overrideEndDate: string = null): CompiledTallyPoint[] {
-  const sortedTallies = tallies.toSorted(cmpTallies);
-  const bounds = determineDateBounds(
-    overrideStartDate, overrideEndDate,
-    sortedTallies[0]?.date, sortedTallies[sortedTallies.length - 1]?.date
-  );
-
-  // first populate the base object and the counts (totals come later)
-  const eachDay = eachDayOfInterval({ start: parseDateString(bounds.start), end: parseDateString(bounds.end) }).map(d => formatDate(d));
-  const compiledPoints: CompiledTallyPoint[] = eachDay.map(date => {
-    const base = {
-      date,
-      count: Object.keys(TALLY_MEASURE_INFO).reduce((obj, measure) => { obj[measure] = 0; return obj; }, {}),
-      total: Object.keys(TALLY_MEASURE_INFO).reduce((obj, measure) => { obj[measure] = 0; return obj; }, {}),
-    };
-
-    const todayTallies = sortedTallies.filter(tally => tally.date === date);
-    for(const tally of todayTallies) {
-      base.count[tally.measure] += tally.count;
-    }
-
-    return base;
-  });
-
-  // now go through and accumulate the totals
-  for(let i = 0; i < compiledPoints.length; ++i) {
-    for(const measure of Object.keys(compiledPoints[i].total)) {
-      compiledPoints[i].total[measure] = compiledPoints[i].count[measure] + (i > 0 ? compiledPoints[i-1].total[measure] : 0);
-    }
-  }
-
-  return compiledPoints;
-}
-function determineDateBounds(
-  overrideStartDate?: string, overrideEndDate?: string,
-  firstTallyDate?: string, lastTallyDate?: string,
-  minimumDaySpan: number = 0
-): { start: string, end: string } {
-  const today = new Date();
-
-  const start = overrideStartDate ?? firstTallyDate ?? formatDate(today);
-  const end = overrideEndDate ?? lastTallyDate ?? formatDate(addDays(parseDateString(start), minimumDaySpan));
-
-  return { start, end };
 }
 
 export function normalizeTallies(tallies: Tally[]): CountedTallyPoint[] {