+++ src/components/goal/TargetStats.vue
@@ -6,11 +6,10 @@ import type { Goal } from 'src/lib/api/goal.ts';
 import { GoalTargetParameters } from 'server/lib/models/goal.ts';
 import type { Tally } from 'src/lib/api/tally.ts';
 import { TALLY_MEASURE } from 'server/lib/models/tally.ts';
-import { TALLY_MEASURE_INFO } from 'src/lib/tally.ts';
+import { TALLY_MEASURE_INFO, compileTallies } from 'src/lib/tally.ts';
 
 import StatTile from './StatTile.vue';
 import { formatDuration, parseDateString } from 'src/lib/date.ts';
-import { compileTallies } from '../chart/chart-functions.ts';
 import { commaify } from 'src/lib/number.ts';
 
 const props = defineProps<{